+++ .gitignore
@@ -15,3 +15,4 @@ my_settings_local.py
 **/*.egg-info
 .coverage
 testsettings2.py
+*.bat
